PhotoSlideShow/requirements.txt
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
Flask==3.0.2
Flask-JWT-Extended==4.6.0
gunicorn==21.2.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
moviepy==1.0.3
mysql-connector-python==8.3.0
numpy==1.26.4
packaging==24.0
pillow==10.2.0
proglog==0.1.10
psycopg2==2.9.9
PyJWT==2.8.0
requests==2.31.0
tqdm==4.66.2
urllib3==2.2.1
Werkzeug==3.0.1
